refactor: consolidate fetchWithTimeout into shared utility
