+++ src/utils/fetch-timeout.ts
@@ -1,0 +1,24 @@
+/**
+ * Fetch wrapper that adds timeout support via AbortController.
+ *
+ * @param url - The URL to fetch
+ * @param init - RequestInit options (headers, method, body, etc.)
+ * @param timeoutMs - Timeout in milliseconds
+ * @param fetchFn - The fetch implementation to use (defaults to global fetch)
+ * @returns The fetch Response
+ * @throws AbortError if the request times out
+ */
+export async function fetchWithTimeout(
+  url: string,
+  init: RequestInit,
+  timeoutMs: number,
+  fetchFn: typeof fetch = fetch,
+): Promise<Response> {
+  const controller = new AbortController();
+  const timer = setTimeout(() => controller.abort(), Math.max(1, timeoutMs));
+  try {
+    return await fetchFn(url, { ...init, signal: controller.signal });
+  } finally {
+    clearTimeout(timer);
+  }
+}
